Move service file to core package
